Dungtvhe140366 v1

See merge request zuyfun/vn-edu-fpt-capstone!68

+++ src/main/java/vn/edu/fpt/capstone/service/impl/RoomServiceImpl.java
@@ -105,12 +105,12 @@ public class RoomServiceImpl implements RoomService {
 	}
 
 	@Override
-	public int minArea(Long idHouse) {
+	public float minArea(Long idHouse) {
 		return roomRepository.getMinArea(idHouse);
 	}
 
 	@Override
-	public int maxArea(Long idHouse) {
+	public float maxArea(Long idHouse) {
 		return roomRepository.getMaxArea(idHouse);
 	}
